requirements.txt
# Core dependencies
python-binance>=1.0.19
httpx>=0.27.0
websockets>=12.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0
yfinance>=0.2.36

# Machine learning
scikit-learn>=1.4.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# CLI
typer[all]>=0.9.0
rich>=13.0.0

# Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
